+++ src/utils/date.js
@@ -6,7 +6,7 @@
  * comparison, formatting, and static factories.
  *
  * @file       Date core class
- * @version    2.1.11
+ * @version    2.2.0
  * @license    MIT
  */